+++ vortex-layout/src/layouts/zoned/mod.rs
@@ -1,15 +1,28 @@
+//! Zoned layouts wrap a data layout with an auxiliary per-zone statistics layout.
+//!
+//! The zoned layout tree has exactly two children:
+//! - a transparent `data` child containing the underlying column data
+//! - an auxiliary `zones` child containing one row of aggregate statistics per zone
+//!
+//! Metadata stores the logical zone length in rows plus the sorted list of statistics present in
+//! the auxiliary table. During scans, pruning first evaluates a falsification predicate against
+//! the `zones` child and only forwards surviving rows to the underlying `data` child.
+
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 mod builder;
+mod pruning;
 mod reader;
+mod schema;
 pub mod writer;
 pub mod zone_map;
 
 use std::sync::Arc;
 
-pub use builder::MAX_IS_TRUNCATED;
-pub use builder::MIN_IS_TRUNCATED;
+pub(crate) use builder::StatsAccumulator;
+pub use schema::MAX_IS_TRUNCATED;
+pub use schema::MIN_IS_TRUNCATED;
 use vortex_array::DeserializeMetadata;
 use vortex_array::SerializeMetadata;
 use vortex_array::dtype::DType;
@@ -35,7 +48,7 @@ use crate::VTable;
 use crate::children::LayoutChildren;
 use crate::children::OwnedLayoutChildren;
 use crate::layouts::zoned::reader::ZonedReader;
-use crate::layouts::zoned::zone_map::ZoneMap;
+use crate::layouts::zoned::schema::stats_table_dtype;
 use crate::segments::SegmentId;
 use crate::segments::SegmentSource;
 use crate::vtable;
@@ -48,7 +61,8 @@ impl VTable for Zoned {
     type Metadata = ZonedMetadata;
 
     fn id(_encoding: &Self::Encoding) -> LayoutId {
-        LayoutId::new("vortex.stats") // For legacy reasons, this is called stats
+        // For legacy reasons the serialized layout encoding ID is still `vortex.stats`.
+        LayoutId::new("vortex.stats")
     }
 
     fn encoding(_layout: &Self::Layout) -> LayoutEncodingRef {
@@ -81,10 +95,9 @@ impl VTable for Zoned {
     fn child(layout: &Self::Layout, idx: usize) -> VortexResult<LayoutRef> {
         match idx {
             0 => layout.children.child(0, layout.dtype()),
-            1 => layout.children.child(
-                1,
-                &ZoneMap::dtype_for_stats_table(layout.dtype(), &layout.present_stats),
-            ),
+            1 => layout
+                .children
+                .child(1, &stats_table_dtype(layout.dtype(), &layout.present_stats)),
             _ => vortex_bail!("Invalid child index: {}", idx),
         }
     }
@@ -145,12 +158,15 @@ impl VTable for Zoned {
     }
 }
 
+/// Encoding marker for the zoned layout.
 #[derive(Debug)]
 pub struct ZonedLayoutEncoding;
 
-/// Annotates a data layout with per-zone aggregate statistics (e.g. min, max, null count).
+/// A layout that annotates a data child with one row of aggregate statistics per zone.
 ///
-/// During reads, zone maps allow entire zones to be skipped when a filter predicate cannot match.
+/// The first child is the underlying data layout. The second child is an auxiliary stats table
+/// whose rows align with logical row zones of length `zone_len`, except for the final partial zone.
+/// During reads, pruning uses the stats table to skip zones whose rows cannot satisfy a filter.
 #[derive(Clone, Debug)]
 pub struct ZonedLayout {
     dtype: DType,
@@ -169,7 +185,7 @@ impl ZonedLayout {
         if zone_len == 0 {
             vortex_panic!("Zone length must be greater than 0");
         }
-        let expected_dtype = ZoneMap::dtype_for_stats_table(data.dtype(), &present_stats);
+        let expected_dtype = stats_table_dtype(data.dtype(), &present_stats);
         if zones.dtype() != &expected_dtype {
             vortex_panic!("Invalid zone map layout: zones dtype does not match expected dtype");
         }
@@ -191,6 +207,10 @@ impl ZonedLayout {
     }
 }
 
+/// Serialized zoned-layout metadata.
+///
+/// `zone_len` is the logical row length of each zone. `present_stats` is the sorted list of
+/// statistics stored in the auxiliary stats-table child.
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct ZonedMetadata {
     pub(super) zone_len: u32,
@@ -317,7 +337,7 @@ mod tests {
             FlatLayout::new(0, dtype.clone(), SegmentId::from(0), read_ctx.clone()).into_layout(),
             FlatLayout::new(
                 0,
-                ZoneMap::dtype_for_stats_table(&dtype, &[]),
+                stats_table_dtype(&dtype, &[]),
                 SegmentId::from(1),
                 read_ctx,
             )
